Stage 1: extract pure leaf helpers into modules
Move side-effect-free helpers out of the views god-module:
- apply_highlight_window -> solr/postprocess.py
- parse_host -> solr/transport.py
- extract_docs_values -> solr/bigquery.py
- safe_int, ClosingTuple -> solr/_compat.py

views.py delegates to these via thin instance methods so existing call sites
and solr.views.* patch targets keep resolving. Behavior unchanged: 48 passing.
